pyproject.toml
[tool.poetry]
name = "visionify-backend"
version = "0.1.0"
description = ""
authors = ["ShyamSundhar1411 <clashwithchiefrpjyt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
fastapi = {extras = ["all"], version = "^0.108.0"}
uvicorn = {extras = ["standard"], version = "^0.25.0"}
joblib = "^1.3.2"
pillow = "^10.1.0"
python-multipart = "^0.0.6"
hypercorn = "^0.17.3"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
pre-commit = "^4.0.0"
bandit = "^1.7.10"
pylint = "^3.3.1"
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"

[tool.flake8]
max-line-length = 88
max-complexity = 12
ignore = ["E402","E501","W503","D100","C901","B008"]
exclude = "__init__.py"


[tool.pylint.MESSAGE_CONTROL]
disable = """
    too-many-locals,
    line-too-long,
    abstract-method,
    no-member,
    missing-module-docstring,
    missing-class-docstring,
    too-many-ancestors,
    too-many-instance-attributes,
    too-few-public-methods,
    too-many-arguments,
    too-many-branches,
    missing-function-docstring,
    broad-exception-caught,
    protected-access,
    duplicate-code,
    return-in-init,
    wildcard-import,
    too-many-lines,
    no-name-in-module,
    too-many-return-statements,
    bad-except-order
"""
enable = "useless-suppression"

[tool.pylint.REPORTS]
reports="no"

[tool.pylint.FORMAT]
max-line-length="88"

[tool.pylint.BASIC]
good-names = "_,i,e,setUp,tearDown,maxDiff"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
default_section = "THIRDPARTY"
known_first_party = []
known_third_party = ["PIL", "fastapi", "numpy", "pydantic_settings", "tensorflow"]
[tool.black]
line-length = 88
[tool.bandit]
exclude_dirs = ["tests"]
tests = ["B201", "B301"]
skips = ["B101", "B601"]

[tool.pytest]
python_files = ["test_*.py", "*_test.py", "tests/*/*.py"]

[tool.mypy]
files = "backend"
ignore_missing_imports = true
strict = true
disallow_untyped_calls = true
disallow_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
plugins = [
    "fastapi",
]

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
